@@ -44,7 +44,7 @@ class App extends Component {
   }
 
   handleDelete = counterId => {
-    const counters = this.state.counters.filter(c => c.id != counterId);
+    const counters = this.state.counters.filter(c => c.id !== counterId);
     this.setState({ counters });
   }
 }
